Fix following error on cygwin.
make.c:196:10: error: array subscript has type `char' [-Werror=char-subscripts]
          *p = tolower(*p);
          ^

+++ src/make.c
@@ -193,7 +193,7 @@ static char *make_elab_name(tree_t t)
       if (*p == '.')
          *p = '\0';
       else
-         *p = tolower(*p);
+         *p = tolower((int)*p);
    }
 
    return name;
